Splitted out the position controller.
