+++ src/main/resources/application/sequence.fxml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.Label?>
-<fx:root type="javafx.scene.control.Label" xmlns:fx="http://javafx.com/fxml"/>
+<fx:root type="javafx.scene.control.Label" xmlns:fx="http://javafx.com/fxml" styleClass="sequence"/>
